All tests 'pass' although checks for warnings and optimisations haven't been implemented

+++ SplCompiler/Compiler Testing Scripts/Tests/Test156.c
@@ -13,7 +13,7 @@ void spl_Test()
 {
 	double spl_a = 1;
 
-	int _spl_real_by = 0.100000;
+	double _spl_real_by = 0.100000;
 	for (spl_a = 0.100000; _spl_real_by < 0 ? spl_a >= 0.500000 : spl_a <= 0.500000; spl_a += _spl_real_by)
 	{
 		printf("%lf", spl_a);
